gradle conflict fix and File structure refactoring
