--- Spring/src/main/java/GFO/Spring/controller/ApiResponses.java
@@ -1,2 +1,0 @@
-package GFO.Spring.controller;
-//커밋 실수로 주석한번 답니다
