+++ Spring/src/main/java/GFO/Spring/domain/email/service/EmailService.java
@@ -1,4 +1,4 @@
-package GFO.Spring.service;
+package GFO.Spring.domain.email.service;
 
 public interface EmailService {
     String sendSimpleMessage(String to)throws Exception;
