+++ Spring/src/main/java/GFO/Spring/global/configuration/email/EmailConfig.java
@@ -1,4 +1,4 @@
-package GFO.Spring.configuration;
+package GFO.Spring.global.configuration.email;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
@@ -9,7 +9,6 @@ import org.springframework.mail.javamail.JavaMailSenderImpl;
 @Configuration
 @RequiredArgsConstructor
 class EmailConfig {
-    //커밋 실수로 주석한번 답니다.
     private final MailProperties mailProperties;
 
     @Bean
